requirements.txt
FPDF==1.7.2
Pillow==9.5.0
tweepy==4.14.0